+++ styles/nsp/info.js
@@ -1,8 +1,9 @@
 export default {
   name: "nsp",
   description: "Nintendo Switch Pro Controller",
-  main: "pro_controller_gray.svg",
+  main: "pro_controller_dark.svg",
   skins: {
+    dark: "pro_controller_dark.svg",
     gray: "pro_controller_gray.svg",
   },
 };