+++ styles/ps4/info.js
@@ -1,0 +1,8 @@
+export default {
+  name: "ps4",
+  description: "Playstation 4 Controller",
+  main: "ps4_gray.svg",
+  skins: {
+    gray: "ps4_gray.svg",
+  },
+};