+++ styles/style_list.js
@@ -1,1 +1,1 @@
-export default ["gcn", "n64", "nsp", "wcc"];
+export default ["gcn", "n64", "nsp", "ps4", "wcc"];